Move `bitfield` type from Spicy to HILTI.
This prepares for making `bitfield` a fully-supported HILTI type. This
change only moves existing code over to HILTI without breaking
Spicy-side functionality. There are still HILTI-side pieces missing
for providing full support, which will come next.
