+++ hilti/toolchain/include/ast/types/all.h
@@ -5,6 +5,7 @@
 #include <hilti/ast/types/address.h>
 #include <hilti/ast/types/any.h>
 #include <hilti/ast/types/auto.h>
+#include <hilti/ast/types/bitfield.h>
 #include <hilti/ast/types/bool.h>
 #include <hilti/ast/types/bytes.h>
 #include <hilti/ast/types/doc-only.h>
