+++ hilti/toolchain/src/ast/types/bitfield.cc
@@ -1,10 +1,9 @@
 // Copyright (c) 2020-2023 by the Zeek Project. See LICENSE for details.
 
+#include <hilti/ast/types/bitfield.h>
 #include <hilti/ast/types/tuple.h>
 
-#include <spicy/ast/types/bitfield.h>
-
-using namespace spicy;
+using namespace hilti;
 
 hilti::optional_ref<const type::bitfield::Bits> type::Bitfield::bits(const ID& id) const {
     for ( const auto& b : bits() ) {
