+++ hilti/toolchain/src/compiler/parser/scanner.ll
@@ -56,7 +56,7 @@ hexit     [0-9a-fA-F]
 hexs      {hexit}+
 E         ([Ee][+-]?{digits})
 P         ([Pp][+-]?{digits})
-decfloat  {digits}{E}|{digit}*\.{digits}{E}?|{digits}\.{digit}*{E}?
+decfloat  {digits}{E}|{digit}*\.{digits}{E}?|{digits}\.{digits}{E}?
 hexfloat  0[xX]({hexit}+{P}|{hexit}*\.{hexit}+{P}?|{hexit}+\.{hexit}*{P}?)
 id        [a-zA-Z_]|[a-zA-Z_][a-zA-Z_0-9-]*[a-zA-Z_0-9]|[$][$]
 property  %[a-zA-Z_][a-zA-Z_0-9-]*
@@ -84,6 +84,7 @@ assert                return token::ASSERT;
 assert-exception      return token::ASSERT_EXCEPTION;
 auto                  return token::AUTO;
 begin                 return token::BEGIN_;
+bitfield              return token::BITFIELD;
 bool                  return token::BOOL;
 break                 return token::BREAK;
 bytes                 return token::BYTES;
