+++ hilti/toolchain/src/compiler/visitors/resolver.cc
@@ -474,6 +474,43 @@ struct Visitor : public visitor::PostOrder<void, Visitor> {
         }
     }
 
+    void operator()(const type::bitfield::Bits& b, position_t p) {
+        if ( type::isResolved(b.itemType()) )
+            return;
+
+        Type t = b.ddType();
+
+        if ( auto a = AttributeSet::find(b.attributes(), "&convert") ) {
+            t = a->valueAsExpression()->get().type();
+            if ( ! type::isResolved(t) )
+                return;
+        }
+
+        logChange(p.node, t);
+        p.node.as<type::bitfield::Bits>().setItemType(t);
+        modified = true;
+    }
+
+    void operator()(const type::Bitfield& b, position_t p) {
+        if ( type::isResolved(b.type()) )
+            return;
+
+        std::vector<hilti::type::tuple::Element> elems;
+
+        for ( const auto& bit : b.bits() ) {
+            if ( ! type::isResolved(bit.itemType()) )
+                return;
+
+            elems.emplace_back(bit.id(), bit.itemType());
+        }
+
+        Type t = type::Tuple(std::move(elems), b.meta());
+        assert(type::isResolved(t));
+        logChange(p.node, t);
+        p.node.as<type::Bitfield>().setType(t);
+        modified = true;
+    }
+
     void operator()(const type::Enum& m, position_t p) {
         if ( type::isResolved(p.node.as<Type>()) )
             return;
